3.3. Malicious registry signing blocks deal participation
We will check that sender matches the signature sent in so that someone can't steal a signature and front run, but we don't want "open deal" agreements to be restricted to finalizers as they are used for other use cases.

If a deal gets grief'd, we would suggest to send a closed offer to the party attempting to invest.